% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comble_posee.R
\name{comble_posee}
\alias{comble_posee}
\title{comble posee avec les eventuelle colonne rajouté, comme si elles avaient été posée a tout le monde}
\usage{
comble_posee(dataset, posee)
}
\arguments{
\item{posee}{}
}
\description{
comble posee avec les eventuelle colonne rajouté, comme si elles avaient été posée a tout le monde
}
