% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Title}
\usage{
compare(dataset, var, tidy = TRUE)
}
\arguments{
\item{tidy}{}
}
\description{
Title
}
\examples{

library(dplyr)
compare(iris,Species)
compare(iris \%>\% rename(Spec_ies=Species),Spec_ies)
compare(iris \%>\% rename(Spec_ies=Species),Spec_ies,tidy=FALSE)


}
