% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_inverse}
\alias{compare_inverse}
\title{compare mais retourne les tableau}
\usage{
compare_inverse(dataset, var, tidy = TRUE)
}
\arguments{
\item{tidy}{}
}
\description{
compare mais retourne les tableau
}
