% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition.R
\name{condition}
\alias{condition}
\title{genrere une table de}
\usage{
condition(dataset, ...)
}
\arguments{
\item{dataset}{un jeu de données}

\item{...}{des conditions}
}
\value{
un data.frame avec des 0 et des 1
}
\description{
genrere une table de
}
\examples{


data("demo_cloudy")
demo_cloudy \%>\%
 condition(
   "tu fumes combien ?" = list(`tu fumes ?` == 'oui' & age > 18)
   , "date des derniere regles" = list(sexe == 'F')
 )


}
