% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dessin_quali.R
\name{dessin_quali_all}
\alias{dessin_quali_all}
\title{Title}
\usage{
dessin_quali_all(dataset, groupe)
}
\arguments{
\item{dataset}{un data.frame}
}
\value{
list of ggplot
}
\description{
Title
}
\examples{
dessin_quali_all(iris)

}
