% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{fetch_data}
\alias{fetch_data}
\title{Fetch USQ Data from AARNET Cloudstor, OneDrive or Sharepoint}
\usage{
fetch_data(url)
}
\arguments{
\item{url}{User provided url generated by "sharing" the file via any of the
supported services}
}
\description{
Fetch USQ Data from AARNET Cloudstor, OneDrive or Sharepoint
}
