% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dessin_quali.R
\name{dessin_quali_var}
\alias{dessin_quali_var}
\title{Title}
\usage{
dessin_quali_var(dataset, var, groupe, camembert = FALSE)
}
\arguments{
\item{var}{}
}
\description{
Title
}
\examples{
dessin_quali_var(iris,Species)
dessin_quali(iris$Species)

dessin_quali(iris$Species,"coucou",camembert = TRUE)
dessin_quali(iris$Species,"coucou",camembert = FALSE)
dessin_quali_var(dataset = iris,var = Species,camembert = TRUE)
dessin_quali_var(dataset = iris,var = Species,camembert = FALSE)
dessin_quali_all(iris)
dessin_quali_all(iris,camembert = TRUE)

}
