% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{fetch_data}
\alias{fetch_data}
\title{Fetch Data from AARNET Cloudstor, OneDrive, Sharepoint, GDrive, GitHub or any Other Web-based Provider}
\usage{
fetch_data(url, which = NULL, file_ext = NULL)
}
\arguments{
\item{url}{User provided url generated by "sharing" the file via any of the
supported filetypes of \code{\link[rio:import]{rio::import()}}, also works with raw GitHub or any
other provider from which a raw data file can be downloaded and imported.
Character.}

\item{which}{User provided spreadhseet for '.xslx' and '.ods' files either
as the number of the sheet in order or the name of the sheet. For .Rdata
objects it can be an object name. See \code{\link[rio:import]{rio::import()}} for further details}

\item{file_ext}{Optional user provided file type extension. If provided,
\code{fetch_data()} will attempt to use this as instruction about the file
type being imported, \emph{e.g.} "xlsx" or "csv". If not provided,
\code{fetch_data()} will do its best to determine the file type automagically and
import the file. Character.}
}
\value{
A `data.frame`` of data derived from any '.csv', '.xls(x)', '.txt',
or '.ods' file provided that \R can import
}
\description{
Given a URL that resolves to a downloadable data file, \code{fetch_data()} will
attempt to download and import the data.
}
